+++ CHANGES.rst
@@ -10,6 +10,8 @@ Unreleased
 - Added ``custom_historical_attrs`` to ``bulk_create_with_history()`` and
   ``bulk_update_with_history()`` for setting additional fields on custom history models
   (gh-1248)
+- Passing an empty list as the ``fields`` argument to ``bulk_update_with_history()`` is
+  now allowed; history records will still be created (gh-1248)
 
 
 3.4.0 (2023-08-18)